Fixed the test harness
Previously, on success this was exiting with a non-numeric
result (an empty list), which python converted into a non-zero
exit status. This, in turn, meant that test success looked like
a failure.

This commit fixes that by changing the exit status to be the
number of failing test cases rather than the list of failing
test cases.