Support Python 3
We had tried to support Python 3 previously, but we were not testing
under Python 3 to ensure that we maintained that support.

This change adds testing for Python 3 and fixes the breakages that have
crept in since our previous effort

This also fixes a transient error in the bigquery tests where
the test only passed if they completed in less than 1 microsecond.