Fetch the application default credentials correctly
After calling `GoogleCredentials.get_application_default()`, the
returned credentials are not guaranteed to be usable until after
calling `.create_scoped(...)` on them with the set of scopes
requested.

There are 4 types of credentials that may be returned when we
call into `GoogleCredentials.get_application_default()`:

0. An instance of `client.GoogleCredentials`
1. An instance of `gce.AppAssertionCredentials`
2. An instance of `appengine.AppAssertionCredentials`
3. An instance of `service_account._JWTAccessCredentials`

For cases 0 and 1, that subsequent call is actually a no-op, so
we did not realize previously that we were missing it.

However, for cases 2 and 3 that subsequent call is required, and
attempting to use the credentials without it will result in error
messages from the server saying that scopes being missing or empty
is not supported.

This change fixes that by setting the scopes to the singleton list
of the 'cloud-platform' scope. That is not ideal, as there are APIs
outside of the Cloud Platform that a Datalab user may want to use.

However, it is a step forward, as it does not impact the scenarios
which already worked (because the `create_scoped` call in those
cases does nothing), but it does allow previously broken scenarios
to work for Cloud Platform APIs.