Only scope credentials if necessary.
The application default credentials are a bit more complicated than
what we were previously supporting. They can be one of multiple
different types, and depending on the type calling `create_scoped`
can either be required or can generate a warning.

To work around this, we have to check the resulting credential's
`create_scoped_required` method, and only create scoped credentials
if that method returns `True`.