Support the new BigQuery types in charts (#68)
BigQuery's new Standard SQL support adds additional data
[types](https://cloud.google.com/bigquery/sql-reference/data-types)
that might occur in the schema field of results.

These new types meant that we would get key errors when looking up
the corresponding type to use for charting the data.

Some of those new types correspond to the set of types supported by
[Google Charts](https://developers.google.com/chart/interactive/docs/reference#DataTable_addColumn)
and some do not.

For the new types that have a correspondence we add that mapping to
the code. For any other types, we simply map them to 'string' (to
avoid a key error).

This change also adds charting support for the new date and times types.