This is a performance fix for loading data into paged table. Currently (#84)
when user clicks next page on a paged table, it downloads ALL data for
that table which makes it very slow on clicking "next page" the first
time. Also it consumes memory expectedly. The fix adds the offset and
count to the query so only the data needed is downloaded.