Remove context from query construction, move it to execution (#180)
The reason for the change is to separate query creation from execution, so that queries can be executed with different settings (project ids, billing tiers... etc), and reduce confusion if the user tries to execute a query that takes subqueries built with different context settings. Instead, query execution is the only context that is taken into account.