Remove the helper methods for bigquery dialect
This change removes the `datalab_bq_dialect` and `set_datalab_bq_dialect`
methods from the new modules. This is because the new modules
only support a single dialect (Standard SQL).

The old module is updated to install its version of those methods
regardless of whether or not the new modules are installed.

This does not remove references to 'bigquery_dialect' from the
`google.datalab.Context` object's config. That is a much larger
change which will be done separately.