Allow colons in table names.
With Standard SQL, the separator between project name and dataset
name was changed from a colon to a period. As such, we changed our
regex to replace periods with colons. However, colons are still
allowed within a project name, so the regex needs to still allow
colons in the project name.