Fix subquery order dependency in Query constructor (#302)
Fixes #282

The problem is subqueries were stored in a dictionary while traversing dependencies, which does not preserve order, so the dependency tree would get flattened in the expanded sql with the order of the hashes of subquery names. Add to this the fact that BigQuery requires a correct dependency order for subqueries, for example the following is rejected:

with q2 as (select * from q1),
       q1 as (select * from some_table)
select * from q2
This PR:

Fixes this by using an list of tuples instead of dictionary.
Adds nice indentation formatting for expanded queries (test modification shows this).
Adds a regression test.