Set up tox, and use it for local testing and Travis.
[tox](https://tox.readthedocs.io/) is a Python test automation tool: it makes
it easy to automate running the tests for your project in several
environments. Under the hood, it just uses virtualenv to manage installed
deps.

If you have python2.7 and python3.5 installed, simply running `tox` at the
root of the repo will now run tests in both versions, as well as `flake8`.

This PR makes two changes:

* first, convert the existing `.travis.yml` into `tox.ini`, setting up three
  tox environments based on the configuration that was already
  there. (Relatedly, I added the `.tox` directory to `.gitignore`).

* Now that #330 is fixed (wooo @brandondutra), I switched the `.travis.yml`
  over to use `tox`. Now there are separate travis environments for each tox
  environment, which has two advantages: (1) these three all run in parallel,
  speeding up builds, and (2) build logs are separated, making it faster to
  figure out what caused a given failure.