Fix a bug in `%bq query` expansion with UDFs.
Currently, attempting to call `%bq query --udfs <some udf object>` will fail
with a stack trace like

```python
TypeErrorTraceback (most recent call last)
<ipython-input-27-0f9364e227db> in <module>()
----> 1 extract_params_query.sql()

/tmp/tmpZ1tEW3/dist-packages/google/datalab/bigquery/_query.pyc in sql(self)
    183   def sql(self):
    184     """ Get the SQL for the query. """
--> 185     return self._expanded_sql()
    186
    187   @property

/tmp/tmpZ1tEW3/dist-packages/google/datalab/bigquery/_query.pyc in _expanded_sql(self, sampling)
    148
    149     if udfs:
--> 150       expanded_sql += '\n'.join([udfs[udf]._expanded_sql() for udf in udfs])
    151       expanded_sql += '\n'
    152

TypeError: list indices must be integers, not tuple
```

The issue here is a mismatch between use of `query._udfs`, which is a list of
`(name, udf)` tuples, and `query.udfs`, which is `dict(query._udfs)`.

This occurs in the
[tutorials](https://github.com/googledatalab/notebooks/blob/master/tutorials/BigQuery/UDFs%20in%20BigQuery.ipynb).

Switching to tuples everywhere fixes this, as verified by the test.