sd package: all categorical columns are multi-label columns (#351)
* all categorical columns are multi-label columns

This means that all string columns are split on space. Also, there
is no difference between categorical columns, multi-label categorical
columns, and text columns.

local/cloud preprocessing updated to build vocabs after splitting. We
also now output the count of each label in a column. This might allow
for tfidf and other weights to be supported.

* flake8

* review comments

* flake8