Follow up on review comments.
The commandParser change is needed so we can display multi-line help for sub commands (%%ml predict in this case).

Also fix python 3 compatibility issue by using a later version of Pillow. The fix is a bit odd (by upgrading the pillow in image classification package), but eventually we will also update Pillow on datalab image in a separate change.