Add an add_cell_argument to CommandParser so cell argument is a first-class thing.
Change the impl so we no longer overriding add_argument(), because that way we will need to also deal with add_exclusive_group() call, which is too hacky. Instead now we find all subparser args by its actions.

It remains true that args added by "add_argument()" can be provided in cell or line. But in one cell input the same values cannot be provided in both.