Add a feature in line/cell magic parsing, so that '--config1 value1' … (#434)
* Add a feature in line/cell magic parsing, so that '--config1 value1' can be specified in cell 'config1: value1'. This makes the notebook input pretty if the line args are too long.

I ran most of the sample notebooks to make sure it does not break existing notebooks, and also verified in these notebooks line param can be moved to cell too except for sql cell.

Note:

1. We do this only on the last subcommand. If your commands are like "%ml subcommand1 --config1 value1 subcommand2 --config2 value2", then it only tries getting "config2" from cell. But in reality we don't have those cases that there are arguments in multiple subcommands.

2. Only param beginning with '--' is supported.

* Add back tests that were removed accidentally.

* Add an add_cell_argument to CommandParser so cell argument is a first-class thing.

Change the impl so we no longer overriding add_argument(), because that way we will need to also deal with add_exclusive_group() call, which is too hacky. Instead now we find all subparser args by its actions.

It remains true that args added by "add_argument()" can be provided in cell or line. But in one cell input the same values cannot be provided in both.

* json indent.

* Fix test.

* Add a comment.

* Follow CR comments.