ML Workbench: If top-n is 0, instead of outputing "predicted, predicted_2,...probability, probability_2,...", output "predicted, class1, class2,...". (#535)
For example, if top-n is 0, before the output was:

predicted: daisy
predicted_2: roses
predicted_3: sunflower
probability: 0.9
probability_2: 0.1
probability_3: 0.1

With this change it will be:
predicted: daisy
daisy: 0.9
roses: 0.1
sunflower: 0.1

The reason for the change is that if all classes are included, top-n results are not that useful. But in some cases, it is convenient to have a column of a certain class's probability. For example, plotting ROC curve requires prob of a certain class.