Change the log level for 'google.auth._default' to ERROR.
With version 1.3.0 of the "google-auth-library-python", the behavior
changed so that it logs a warning if it reads the credentials from
a metadata server but gets an empty project ID from that server.

The resulting log message is erroneous, telling the user to call
`gcloud config set project ...` to silence the warning, when in
reality it never looks in the gcloud config for that scenario.

This causes issues if you run Datalab with the fake metadata server
[here](https://github.com/googledatalab/datalab/blob/master/sources/web/datalab/metadata.ts),
but do not provide a value for the project ID.

This change mitigates that issue by raising the log level of the
offending module to `ERROR` prior to fetching the credentials.

The error is then set back to the previous level in case the
user is using the `google.auth._default` module in their own code.