Fix the usage of `google.datalab.utils.Iterator`.
The Iterator class takes a `retriever` method to which it passes
two arguments. Until now, the various `retriever` methods passed
in have interpreted the second argument differently; some treat
it as a page size, while others rely on it representing the total
number of items iterated over so far.

This change fixes the inconsitency by making every implementation
treat the second argument as the number of items retrieved so far.

For implementations that were relying on that value being the
page size, they have been updated to calculate the page size themselves
the same way that it would have been calculated by the previous
version of the Iterator class.

This fixes #686