Fix an MLToolbox preprocessing failure caused by pickling of TF objects. (#689)
It seems happening after we upgrade DataFlow version to 2.0. The issue happens when:

You have "import tensorflow as tf" in a DoFn's init.
At run time, DataFlow initializes the DoFn first, and then pickle and unpickle the instance to multiple workers. That means it needs to pickle the state including imports.
Under tensorflow there are some objects that are not pickle friendly. So pickling everything under tensorflow namespace causes it to fail at run time.
The issue also occur if you have global "import tensorflow" statement.

The fix is to move tensorflow import into either DoFn's process(), or start_bundle() function. That way the import occurs for each instance on each worker and no tensorflow pickling is needed.